Updated affix file
Since the stemmer / parser is set to ignore punctuation by default, was running into issues where certain morphemes would not show up after a word that ended in a glottal stop ('). Set those morphemes to occur anywhere (even after vowels), which is not preferred from a spell-check point of view, but works okay when just interpreting parses of existing words.